Format code

+++ modules/organs/action_mysql.php
@@ -8,6 +8,8 @@
  * @Createdate Dec 3, 2010  11:10:39 AM
  */
 
+if (!defined('NV_IS_FILE_MODULES')) die('Stop!!!');
+
 $sql_drop_module = array();
 $sql_drop_module[] = "DROP TABLE IF EXISTS " . $db_config['prefix'] . "_" . $lang . "_" . $module_data . "_rows;";
 $sql_drop_module[] = "DROP TABLE IF EXISTS " . $db_config['prefix'] . "_" . $lang . "_" . $module_data . "_person;";
